Modifications to the sellers table, and also seeds the database with an Admin user for myself.

+++ database/migrations/2024_04_10_132623_create_seller_sites_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(Seller::class)->cascadeOnDelete();
             $table->foreignIdFor(LinkSite::class);
-            $table->decimal('price_guest_post');
-            $table->decimal('price_link_insertion');
+            $table->unsignedTinyInteger('price_guest_post');
+            $table->unsignedTinyInteger('price_link_insertion');
             $table->timestamps();
         });
     }
